+++ test/score-manager-test.js
@@ -70,7 +70,7 @@ buster.testCase('ScoreManager with configuration', {
     'should throw error when plugin does not exist': function () {
       buster.assert.exception(() => {
         this.manager.scoreWith('nonexistent-plugin', {})
-      })
+      }, 'InvalidInputError')
     }
   },
 
@@ -104,7 +104,7 @@ buster.testCase('ScoreManager with configuration', {
       buster.assert.exception(() => scoreManager.create(config))
     },
 
-    'should throw error when plugin was defined without inputs field': function () {
+    'should throw error when plugin was defined without inputs field and inputGroup field': function () {
       let config = {
         plugins: { 'plugin-a': { use: this.stub() } },
         aggregator: this.stubAggregator
@@ -118,9 +118,131 @@ buster.testCase('ScoreManager with configuration', {
         aggregator: this.stubAggregator
       }
       buster.assert.exception(() => scoreManager.create(config))
+    },
+
+    'should throw error when plugin was defined with invalid inputGroup field': function () {
+      let config = {
+        plugins: { 'plugin-a': { use: this.stub(), inputGroup: 'not an array' } },
+        aggregator: this.stubAggregator
+      }
+      buster.assert.exception(() => scoreManager.create(config))
+    },
+
+    'should throw error when plugin was defined with inputGroup field but with invalid content': function () {
+      let config = {
+        plugins: { 'plugin-a': { use: this.stub(), inputGroup: ['invalid content'] } },
+        aggregator: this.stubAggregator
+      }
+      buster.assert.exception(() => scoreManager.create(config))
+    },
+
+    'should throw error when plugin was defined with inputGroup field but with valid and invalid content': function () {
+      let config = {
+        plugins: { 'plugin-a': { use: this.stub(), inputGroup: [['valid'], 'invalid content'] } },
+        aggregator: this.stubAggregator
+      }
+      buster.assert.exception(() => scoreManager.create(config))
+    },
+
+    'should throw error when plugin has no valid use-string': function () {
+      let config = {
+        plugins: { 'plugin-a': { use: 'nonexistent-plugin', inputs: ['a', 'b'] } },
+        aggregator: this.stubAggregator
+      }
+      let manager = scoreManager.create(config)
+      buster.assert.exception(() => manager.score({}), 'InvalidInputError')
     }
   },
+  'ids in mapping objects': {
+    setUp: function () {
+      let aggregator = { combine: this.stub() }
+      this.config = {
+        aggregator,
+        plugins: {}
+      }
+    },
+    'idPath exists with a valid field': function () {
+      this.config.idPath = 'a.b.c[].d.e'
+      buster.refute.exception(() => scoreManager.create(this.config))
+    },
+    'idPath exists but is not a string -> throw error': function () {
+      this.config.idPath = 123
+      buster.assert.exception(() => scoreManager.create(this.config))
+    },
+    'idPath exists but does not contain an': function () {
+      this.config.idPath = 'a.b.c'
+      buster.assert.exception(() => scoreManager.create(this.config))
+    },
+    'idPath exists in the configuration but ends with []': function () {
+      this.config.idPath = 'a[]'
+      buster.assert.exception(() => scoreManager.create(this.config))
+    },
+    'idPath exists but no valid array name': function () {
+      this.config.idPath = '[].c'
+      buster.assert.exception(() => scoreManager.create(this.config))
+    },
+    'idPath exists correct but has more arrays inside': function () {
+      this.config.idPath = 'a.b[].c.d[].e'
+      buster.assert.exception(() => scoreManager.create(this.config))
+    },
+    'iPath does not exist, no error should be given': function () {
+      buster.refute.exception(() => scoreManager.create(this.config))
+    }
+  },
+  'id creation': {
+    setUp: function () {
+      this.stubPluginA = this.stub()
+      this.stubPluginA.returns(0.5)
+      this.stubEval = this.stub()
+      this.stubEval.returns(0.5)
+      this.stubParse = this.stub(aggregatorConfigParser, 'parse').returns({ eval: this.stubEval })
+      this.config = {
+        aggregator: 'aggregator configuration',
+        plugins: {
+          'plugin-a': {
+            use: this.stubPluginA,
+            inputs: ['x', 'y[]']
+          }
+        }
+      }
+    },
+    'return an id if configured correct and id is given': function () {
+      this.config.idPath = 'y[].id'
+      let blob = {
+        x: 'something',
+        y: [
+          { 'id': '123' },
+          { 'id': '321' },
+          { 'id': 'abc' }
+        ]
+      }
 
+      var manager = scoreManager.create(this.config)
+      let score = manager.score(blob)
+      let matcher = [ { id: '123' }, { id: '321' }, { id: 'abc' } ]
+      buster.assert.match(score, matcher)
+    },
+    'appends ids to the mapping objects if they dont exist': function () {
+      this.config.idPath = 'y[].a.id'
+      let blob = {
+        x: 'something',
+        y: [
+          { },
+          { a: { } },
+          { a: { 'id': 'abc' } }
+        ]
+      }
+
+      var manager = scoreManager.create(this.config)
+      let score = manager.score(blob)
+      buster.refute.equals(score[0].id, undefined)
+      buster.refute.equals(score[1].id, undefined)
+      buster.assert.match(score[2], { id: 'abc' })
+      buster.refute.equals(blob.y[0].a.id, undefined)
+      buster.refute.equals(blob.y[1].a.id, undefined)
+      buster.refute.equals(blob.y[2].a.id, undefined)
+    }
+  },
   'score using Aggregator': {
     setUp: function () {
       this.stubPluginA = this.stub()
@@ -149,8 +271,8 @@ buster.testCase('ScoreManager with configuration', {
 
       this.manager.score({ x: {}, y: [0] })
 
-      buster.assert.calledWith(this.stubParse,
-        'aggregator configuration', { 'plugin-a': 0.5, 'plugin-b': 0.8 })
+      buster.assert.calledWith(this.stubParse, 'aggregator configuration')
+      buster.assert.calledWith(this.stubEval, { 'plugin-a': 0.5, 'plugin-b': 0.8 })
     },
 
     'should return the scores returned by the aggregator in field total': function () {
@@ -196,7 +318,8 @@ buster.testCase('ScoreManager with configuration', {
   'plugin failures': {
     setUp: function () {
       this.stubParse = this.stub(aggregatorConfigParser, 'parse')
-      this.stubParse.returns({ eval: this.stub() })
+      this.stubEval = this.stub()
+      this.stubParse.returns({ eval: this.stubEval })
       this.aggregatorSpec = 'not null but not used'
     },
 
@@ -306,7 +429,8 @@ buster.testCase('ScoreManager with configuration', {
 
       manager.score(blob)
 
-      buster.assert.calledWith(this.stubParse, this.aggregatorSpec, {'plugin-a': 1.0})
+      buster.assert.calledWith(this.stubParse, this.aggregatorSpec)
+      buster.assert.calledWith(this.stubEval, {'plugin-a': 1.0})
     }
   }
 })
@@ -458,4 +582,5 @@ buster.testCase('ScoreManager Integration', {
       buster.assert.calledWith(pluginA, { xkey: 'xvalue' }, { ykey: 'yvalue' }, { 'my-special-arg': 100 })
     }
   }
+
 })